+++ scripts/repo/winesapos-build-repo.sh
@@ -75,6 +75,10 @@ makepkg_fn opengamepadui-session-git
 makepkg_fn inputmodule-udev install
 makepkg_fn inputmodule-control
 
+gpg --recv-keys ABAF11C65A2970B130ABE3C479BE3E4300411886
+gpg --recv-keys 647F28654894E3BD457199BE38DBBDC86092693E
+makepkg_fn linux-fsync-nobara-bin
+
 WINESAPOS_REPO_BUILD_LINUX_GIT="${WINESAPOS_REPO_BUILD_LINUX_GIT:-false}"
 if [[ "${WINESAPOS_REPO_BUILD_LINUX_GIT}" == "true" ]]; then
     # Import keys from the two main Linux kernel maintainers: